Rename storageutils to utils, pull repeated methods out into utils file, add imports to List and ListItem component
