+++ src/App.js
@@ -4,7 +4,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './components/Home.jsx';
 import List from './components/List.jsx';
 import AddItem from './components/AddItem.jsx';
-import { getUser } from './storage-utils/storage-utils';
+import { getUser } from './utils/utils';
 
 function App() {
   const userToken = getUser();
