+++ src/components/JoinExistingList.jsx
@@ -1,7 +1,7 @@
 import { db } from '../lib/firebase';
 import { collection, onSnapshot } from '@firebase/firestore';
 import { useState, useEffect } from 'react';
-import { setUser } from '../storage-utils/storage-utils';
+import { setUser } from '../utils/utils';
 import { useNavigate } from 'react-router-dom';
 
 export default function JoinExistingList() {
@@ -32,11 +32,10 @@ export default function JoinExistingList() {
           if (docs.length > 0) {
             setUser(userToken);
             navigate('list');
-          } else {
-            alert(
-              'That list does not exist. Please try again or create a new list.',
-            );
           }
+          alert(
+            'That list does not exist. Please try again or create a new list.',
+          );
         }}
       >
         <label htmlFor="user token">Share Token</label>
