+++ src/components/ListItem.jsx
@@ -1,8 +1,14 @@
 import { useState } from 'react';
 import { Timestamp, doc, updateDoc, deleteDoc } from 'firebase/firestore';
 import { db } from '../lib/firebase';
-import { getUser } from '../storage-utils/storage-utils';
+import { getUser } from '../utils/utils';
 import { calculateEstimate } from '@the-collab-lab/shopping-list-utils';
+import {
+  wasPurchasedWithin24Hours,
+  daysSinceLastPurchase,
+  daysUntilNextPurchase,
+  isActive,
+} from '../utils/utils';
 
 export default function ListItem({ item, index }) {
   const [userToken] = useState(getUser());
@@ -29,37 +35,6 @@ export default function ListItem({ item, index }) {
     });
   };
 
-  const wasPurchasedWithin24Hours = (item) => {
-    let now = Timestamp.now().seconds;
-    let itemPurchaseDate = item.data().lastPurchaseDate;
-    let difference = now - itemPurchaseDate;
-    const secondsInDay = 86400;
-    return difference < secondsInDay;
-  };
-
-  const daysSinceLastPurchase = (item) => {
-    let now = Timestamp.now().seconds;
-    const lastPurchase = item.data().lastPurchaseDate;
-    const itemCreated = item.data().dateItemAdded;
-
-    if (!lastPurchase) {
-      return Math.round((now - itemCreated) / 86400);
-    }
-    return Math.round((now - lastPurchase) / 86400);
-  };
-
-  const daysUntilNextPurchase = (item) => {
-    return item.data().estimatedPurchaseInterval - daysSinceLastPurchase(item);
-  };
-
-  const isActive = (item) => {
-    return (
-      !wasPurchasedWithin24Hours(item) &&
-      item.data().totalPurchases > 0 &&
-      daysSinceLastPurchase(item) < item.data().estimatedPurchaseInterval * 2
-    );
-  };
-
   const determinePurchaseCategory = (item) => {
     if (!isActive(item)) {
       return 'inactive';
@@ -69,9 +44,8 @@ export default function ListItem({ item, index }) {
       return 'soon';
     } else if (daysUntilNextPurchase(item) <= 30) {
       return 'kinda-soon';
-    } else {
-      return 'not-soon';
     }
+    return 'not-soon';
   };
 
   const deleteHandler = (item) => {
